+++ webapp/src/views/UsersList.vue
@@ -18,7 +18,7 @@ import TheHeader from "../components/TheHeader.vue";
 import { CurrentUser } from "../services/LoginService";
 import { UnauthorizedError } from "../utils/Errors";
 import UsersTable from "../components/UsersTable.vue";
-import { swalUserForm } from "../helpers/SwalHelper";
+import { swalConfirm, swalUserForm } from "../helpers/SwalHelper";
 
 const usersService = new UsersService();
 
@@ -47,8 +47,10 @@ export default defineComponent({
       this.users = await usersService.getAllUsers();
     },
     async deleteUser(user: User) {
-      await usersService.deleteUser(user);
-      this.updateUsersList();
+      if (await swalConfirm("This action will permanently delete the user")) {
+        await usersService.deleteUser(user);
+        this.updateUsersList();
+      }
     },
     async editUser(user: User) {
       let newUser = await swalUserForm(user);
